+++ changelog.md
@@ -3,9 +3,11 @@
 - Primitive figure rendering
 - Basic generation
 
-## Version 0.2.1
+## Version 0.2.0
 - Refactorized svg.rs to optimize the rendering
 - Created the latex drawer
 - Completed the documentation as of now.
 - Added a Command Line Interface
 - Pretty error printing
+- Added the .json drawer
+- Added the raw drawer